requirements.txt
kivy==2.2.1
kivymd==1.0.1
pycryptodome==3.20.0
flask==2.3.3
flask-jwt-extended==4.6.0
websockets==12.0
requests==2.31.0
flask-dance
SQLAlchemy==2.0.23
